Crate zeitstempel

zeitstempel is German for “timestamp”.

Time’s hard. Correct time is near impossible.

This crate has one purpose: give me a timestamp as an integer, coming from a monotonic clock source, guaranteed with or without time across suspend/hibernation of the host machine, and let me compare it to other timestamps.

It becomes the developer’s responsibility to only compare timestamps obtained from this clock source. Timestamps are not comparable across operating system reboots.

Why not std::time::Instant?

std::time::Instant fulfills some of our requirements:

It’s monotonic, guaranteed (sort of).

It can be compared to other timespans.

However:

It can’t be serialized.

It’s not guaranteed that the clock source it uses contains (or exclude) suspend/hibernation time across all operating systems.

Example with suspend time

let start = zeitstempel::now();
thread::sleep(Duration::from_millis(2));

let diff = Duration::from_nanos(zeitstempel::now() - start);
assert!(diff >= Duration::from_millis(2));

Example with only awake time

let start = zeitstempel::now_awake();
thread::sleep(Duration::from_millis(2));

let diff = Duration::from_nanos(zeitstempel::now_awake() - start);
assert!(diff >= Duration::from_millis(2));

Supported operating systems

We support the following operating systems:

Windows*

macOS

Linux

Android

iOS

FreeBSD

OpenBSD

For other operating systems there’s a fallback to std::time::Instant, compared against a process-global fixed reference point. We don’t guarantee that measured time includes time the system spends in sleep or hibernation.

* To use native Windows 10 functionality enable the win10plus feature. Otherwise it will use the fallback.

Modules

unix 🔒

Functions

now

Returns a timestamp corresponding to “now”, including suspend time.

now_awake

Returns a timestamp corresponding to “now”, excluding suspend time.