zeitstempel::unix

Function now_awake

pub fn now_awake() -> u64

The time from a clock that cannot be set and represents monotonic time since some unspecified starting point, but that does not increment while the system is asleep.

See clock_gettime and FreeBSD clock_gettime.